Replace energy threshold inputs with calibrated step selects
Options run from 7.5% (10 mins) to 100% (30 mins) using measured
energy steps. Values stored as float. Labels show "X% - Y mins".

+++ ui/app.py
@@ -37,10 +37,10 @@ def save():
 
     c["aggravated_crimes"]["enabled"] = data.get("crimes_enabled", False)
     c["aggravated_crimes"]["primary"]["crime"] = data.get("primary_crime", "pickpocket")
-    c["aggravated_crimes"]["primary"]["energy_threshold"] = int(data.get("primary_threshold", 50))
+    c["aggravated_crimes"]["primary"]["energy_threshold"] = float(data.get("primary_threshold", 50))
     c["aggravated_crimes"].setdefault("away_crime", {})
     c["aggravated_crimes"]["away_crime"]["crime"] = data.get("away_crime", "pickpocket")
-    c["aggravated_crimes"]["away_crime"]["energy_threshold"] = int(data.get("away_threshold", 50))
+    c["aggravated_crimes"]["away_crime"]["energy_threshold"] = float(data.get("away_threshold", 50))
     c["aggravated_crimes"].setdefault("armed", {})
     c["aggravated_crimes"]["armed"]["agg_private"] = data.get("armed_agg_private", False)
     c["aggravated_crimes"]["armed"]["agg_drug_house"] = data.get("armed_agg_drug_house", False)
